System.Metrics.Protocol.Type

Description

The type of the EKG forwarding/accepting protocol.

Since we are using a typed protocol framework this is in some sense the definition of the protocol: what is allowed and what is not allowed.

Documentation

data EKGForward req resp where

A kind to identify our protocol, and the types of the states in the state transition diagram of the protocol.

IMPORTANT NOTE: the following terminology is used:

From the protocol's point of view, two peers talk to each other: the forwarder and the acceptor.

The forwarder is an application that collects EKG metrics and sends them to the acceptor by request.

The acceptor is an application that receives EKG metrics from the forwarder.

You can this of the forwarder as a client, and the acceptor as a server:

1. The client is "initially active side", because it establishes the connection with the server.

2. The server is "initially passive side", because it accepts the connection from the client.

But after the connection is established, the roles are REVERSED: the acceptor becomes an active side because it asks the metrics, the forwarder becomes a passive side because it waits for the request from the acceptor, collects the metrics and sends them to the acceptor.

Constructors

StIdle :: EKGForward req resp
Both acceptor and forwarder are in idle state. The acceptor can send a request for the metrics and the forwarder is waiting for a request.

StBusy :: EKGForward req resp
The acceptor has sent a next request for the metrics. The acceptor is now waiting for a response, and the forwarder is busy getting ready to send a response with new metrics. It is assumed that the forwarder can send a reply immediately (with the current list of the metrics).

StDone :: EKGForward req resp
Both the acceptor and forwarder are in the terminal state. They're done.

data family Message ps (st :: ps) (st' :: ps)

data SingEKGForward (st :: EKGForward req resp) where

Singleton type of EKGForward. Same as:

type SingEKGForward :: EKGForward req resp -> Type
type SingEKGForward = TypeRep

Constructors

SingIdle :: SingEKGForward 'StIdle

SingBusy :: SingEKGForward 'StBusy

SingDone :: SingEKGForward 'StDone